Style: MyOrderItem 컴포넌트화
