+++ src/components/index.js
@@ -1,5 +1,6 @@
-import Pagination from './Pagination';
-import Star from './Star';
-import Modal from './Modal';
-import Carousel from './Carousel';
-export { Pagination, Star, Modal, Carousel };
+import Pagination from "./Pagination";
+import Star from "./Star";
+import Modal from "./Modal";
+import Carousel from "./Carousel";
+import MyOrderItem from "./MyOrderItem";
+export { Pagination, Star, Modal, Carousel, MyOrderItem };
